working version (final, really 200%)

+++ contact.h
@@ -4,6 +4,6 @@
 #include "bf_ispc.h"
 
 /* calculate distances */
-void contact_detection (unsigned long long int s1, unsigned long long int e1, unsigned long long int s2, unsigned long long int e2,  unsigned long long int size, iREAL *t[3][3], iREAL *p[3], iREAL *q[3], iREAL *distance);
+void contact_detection (unsigned int s1, unsigned int e1, unsigned int s2, unsigned int e2,  unsigned int size, iREAL *t[3][3], iREAL *p[3], iREAL *q[3], iREAL *distance);
 
 #endif
